+++ pkg/core/src/tests/GClaimManager.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.11;
 
-import { ERC20 } from "@rari-capital/solmate/src/erc20/ERC20.sol";
+import { ERC20 } from "@rari-capital/solmate/src/tokens/ERC20.sol";
 import { FixedMath } from "../external/FixedMath.sol";
 
 import { Errors } from "@sense-finance/v1-utils/src/libs/Errors.sol";
@@ -21,7 +21,7 @@ contract GClaimsManager is TestHelper {
     /* ========== join() tests ========== */
 
     function testFuzzCantJoinIfInvalidMaturity(uint128 balance) public {
-        uint48 maturity = uint48(block.timestamp - 1 days);
+        uint256 maturity = block.timestamp - 1 days;
         //        uint256 balance = 1e18;
         try alice.doJoin(address(adapter), maturity, balance) {
             fail();
@@ -31,7 +31,7 @@ contract GClaimsManager is TestHelper {
     }
 
     function testFuzzCantJoinIfSeriesDoesntExists(uint128 balance) public {
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         //        uint256 balance = 10e18;
         try alice.doJoin(address(adapter), maturity, balance) {
             fail();
@@ -41,7 +41,7 @@ contract GClaimsManager is TestHelper {
     }
 
     function testFuzzCantJoinIfNotEnoughClaim(uint128 balance) public {
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         (, address claim) = sponsorSampleSeries(address(alice), maturity);
         if (calculateAmountToIssue(balance) == 0) return;
         hevm.warp(block.timestamp + 1 days);
@@ -54,7 +54,7 @@ contract GClaimsManager is TestHelper {
     }
 
     function testFuzzCantJoinIfNotEnoughClaimAllowance(uint128 balance) public {
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         (, address claim) = sponsorSampleSeries(address(alice), maturity);
         if (calculateAmountToIssue(balance) == 0) return;
         hevm.warp(block.timestamp + 1 days);
@@ -69,7 +69,7 @@ contract GClaimsManager is TestHelper {
 
     function testCantJoinAfterFirstGClaimNotEnoughTargetBalance() public {
         adapter.setScale(0.1e18); // freeze scale so no excess is generated
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         (, address claim) = sponsorSampleSeries(address(alice), maturity);
 
         // bob issues and joins
@@ -116,7 +116,7 @@ contract GClaimsManager is TestHelper {
     //     poolManager.setIsTrusted(address(periphery), true);
     //     alice.doApprove(address(stake), address(periphery));
 
-    //     uint48 maturity = getValidMaturity(2021, 10);
+    //     uint256 maturity = getValidMaturity(2021, 10);
     //     (, address claim) = sponsorSampleSeries(address(alice), maturity);
     //     if (calculateAmountToIssue(balance) == 0) return;
 
@@ -193,7 +193,7 @@ contract GClaimsManager is TestHelper {
     /* ========== exit() tests ========== */
 
     function testFuzzCantExitIfSeriesDoesntExists(uint128 balance) public {
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         //        uint256 balance = 1e18;
         try alice.doExit(address(adapter), maturity, balance) {
             fail();
@@ -218,7 +218,7 @@ contract GClaimsManager is TestHelper {
         poolManager.setIsTrusted(address(periphery), true);
         alice.doApprove(address(stake), address(periphery));
 
-        uint48 maturity = getValidMaturity(2021, 10);
+        uint256 maturity = getValidMaturity(2021, 10);
         (, address claim) = sponsorSampleSeries(address(alice), maturity);
         if (calculateAmountToIssue(balance) == 0) return;
 
@@ -240,7 +240,7 @@ contract GClaimsManager is TestHelper {
     // TODO: re-add this test once we use glcaims again
     // function testFuzzExitGClaimWithCollected(uint128 balance) public {
     //     balance = fuzzWithBounds(balance, 1e12);
-    //     uint48 maturity = getValidMaturity(2021, 10);
+    //     uint256 maturity = getValidMaturity(2021, 10);
     //     (, address claim) = sponsorSampleSeries(address(alice), maturity);
     //     // avoid fuzz tests in which nothing is issued
     //     if (calculateAmountToIssue(balance) == 0) return;
