+++ pkg/core/src/tests/WstETHAdapter.tm.sol
@@ -2,7 +2,8 @@
 pragma solidity 0.8.11;
 
 import { FixedMath } from "../external/FixedMath.sol";
-import { SafeERC20, ERC20 } from "@rari-capital/solmate/src/erc20/SafeERC20.sol";
+import { ERC20 } from "@rari-capital/solmate/src/tokens/ERC20.sol";
+import { SafeTransferLib } from "@rari-capital/solmate/src/utils/SafeTransferLib.sol";
 
 // Internal references
 import { Periphery } from "../Periphery.sol";
@@ -56,8 +57,8 @@ contract WstETHAdapterTestHelper is LiquidityHelper, DSTest {
 
     uint64 public constant ISSUANCE_FEE = 0.01e18;
     uint256 public constant STAKE_SIZE = 1e18;
-    uint48 public constant MIN_MATURITY = 2 weeks;
-    uint48 public constant MAX_MATURITY = 14 weeks;
+    uint256 public constant MIN_MATURITY = 2 weeks;
+    uint256 public constant MAX_MATURITY = 14 weeks;
 
     function setUp() public {
         address[] memory assets = new address[](1);
